Retype DataModel response, command, and query models from long to int

+++ src/config/datamodel/EdFi.DmsConfigurationService.DataModel/Model/ApiClient/ApiClientInsertCommand.cs
@@ -9,10 +9,10 @@ namespace EdFi.DmsConfigurationService.DataModel.Model.ApiClient;
 
 public class ApiClientInsertCommand
 {
-    public required long ApplicationId { get; set; }
+    public required int ApplicationId { get; set; }
     public required string Name { get; set; } = "";
     public required bool IsApproved { get; set; }
-    public long[] DataStoreIds { get; set; } = [];
+    public int[] DataStoreIds { get; set; } = [];
 
     public class Validator : AbstractValidator<ApiClientInsertCommand>
     {
