+++ src/config/datamodel/EdFi.DmsConfigurationService.DataModel/Model/ApiClient/ApiClientResponse.cs
@@ -7,11 +7,11 @@ namespace EdFi.DmsConfigurationService.DataModel.Model.ApiClient;
 
 public class ApiClientResponse
 {
-    public long Id { get; set; }
-    public long ApplicationId { get; set; }
+    public int Id { get; set; }
+    public int ApplicationId { get; set; }
     public required string ClientId { get; set; }
     public required Guid ClientUuid { get; set; }
     public required string Name { get; set; }
     public required bool IsApproved { get; set; }
-    public List<long> DataStoreIds { get; set; } = [];
+    public List<int> DataStoreIds { get; set; } = [];
 }
